Addresses issues from code review
